Replace wignerH with wignerH2
