+++ spherical/recursions/wignerH.py
@@ -1,134 +1,148 @@
-import numpy as np
-from functools import lru_cache
-from scipy.special import factorial
-import spherical as sf
-from .. import jit
-
-
-sqrt3 = np.sqrt(3)
-sqrt2 = np.sqrt(2)
-
-
 """Algorithm for computing H, as given by arxiv:1403.7698
 
-dₗⁿᵐ = ϵₙ ϵ₋ₘ Hₗⁿᵐ, where
+H is related to Wigner's (small) d via
 
-       ⎧ 1 for k≤0
-  ϵₖ = ⎨
-       ⎩ (-1)ᵏ for k>0
+    dₗⁿᵐ = ϵₙ ϵ₋ₘ Hₗⁿᵐ,
 
-Symmetry relations:
-H^{m', m}_n(\beta) = H^{m, m'}_n(\beta)
-H^{m', m}_n(\beta) = H^{-m', -m}_n(\beta)
-H^{m', m}_n(\beta) = (-1)^{n+m+m'} H^{-m', m}_n(\pi - \beta)
-H^{m', m}_n(\beta) = (-1)^{m+m'} H^{m', m}_n(-\beta)
+where
 
-1: If n=0 set H_{0}^{0,0}=1.
+         ⎧ 1 for k≤0
+    ϵₖ = ⎨
+         ⎩ (-1)ᵏ for k>0
 
-2: Compute values H^{0,m}_{n}(β)for m=0,...,n and H^{0,m}_{n+1}(β) for m=0,...,n+1 using Eq. (32):
-     H^{0,m}_{n}(β) = (-1)^m \sqrt{(n-|m|)!/(n+|m|)!} P^{|m|}_{n}(cos β)
+H has various advantages over d, including the fact that it can be efficiently
+and robustly valculated via recurrence relations, and the following symmetry
+relations:
 
-3: Use relation (41) to compute H^{1,m}_{n}(β) for m=1,...,n.  Using symmetry and shift of the
-   indices this relation can be written as
-     b^{0}_{n+1} H^{1,m}_{n} = \frac{b^{−m−1}_{n+1} (1−cosβ)}{2} H^{0,m+1}_{n+1}
-                             − \frac{b^{m−1}_{n+1}  (1+cosβ)}{2} H^{0,m−1}_{n+1}
-                             − a^{m}_{n} sinβ H^{0,m}_{n+1}
+    H^{m', m}_n(β) = H^{m, m'}_n(β)
+    H^{m', m}_n(β) = H^{-m', -m}_n(β)
+    H^{m', m}_n(β) = (-1)^{n+m+m'} H^{-m', m}_n(π - β)
+    H^{m', m}_n(β) = (-1)^{m+m'} H^{m', m}_n(-β)
 
-4: Recursively compute H^{m′+1,m}_{n}(β) for m′=1,...,n−1, m=m′,...,n using relation (50) resolved
-   with respect to H^{m′+1,m}_{n}:
-     d^{m′}_{n} H^{m′+1,m}_{n} = d^{m′−1}_{n} H^{m′−1,m}_{n}
-                               − d^{m−1}_{n} H^{m′,m−1}_{n}
-                               + d^{m}_{n} H^{m′,m+1}_{n}
-   (where the last term drops out for m=n).
-
-5. Recursively compute H^{m′−1,m}_{n}(β) for m′=−1,...,−n+1, m=−m′,...,n using relation (50)
-   resolved with respect to H^{m′−1,m}_{n}:
-     d^{m′−1}_{n} H^{m′−1,m}_{n} = d^{m′}_{n} H^{m′+1,m}_{n}
-                                 + d^{m−1}_{n} H^{m′,m−1}_{n}
-                                 − d^{m}_{n} H^{m′,m+1}_{n}
-   (where the last term drops out for m=n).
-
-6: Apply the first and the second symmetry relations above to obtain all other values H^{m′,m}_{n}
-   outside the computational triangle m=0,...,n, m′=−m,...,m.
+Because of these symmetries, we only need to evaluate at most 1/4 of all the
+elements.
 
 """
 
+import numpy as np
+from functools import lru_cache
+from scipy.special import factorial
+import sympy
+
+from .. import jit, LMpM_total_size, LMpM_index
+
+sqrt3 = np.sqrt(3)
+sqrt2 = np.sqrt(2)
+inverse_sqrt2 = 1.0 / sqrt2
+
 
 @jit
-def wedge_size(ℓₘₐₓ):
-    return (ℓₘₐₓ+1) * (ℓₘₐₓ+2) * (2*ℓₘₐₓ+3) // 6
+def ϵ(m):
+    if m <= 0:
+        return 1
+    elif m%2:
+        return -1
+    else:
+        return 1
+
 
 @jit
-def wedge_index(ℓ, mp, m):
+def wedge_size(ℓₘₐₓ, mpₘₐₓ=None):
+    """Total size of wedges of width mpₘₐₓ up to ℓₘₐₓ
+
+    Parameters
+    ----------
+    ℓₘₐₓ : int
+    mpₘₐₓ : int, optional
+        If None, it is assumed to be at least ℓ
+
+    See Also
+    --------
+    wedge_index : Index inside these wedges
+
+    Notes
+    -----
+    Here, it is assumed that only data with m≥|m'| are stored, and only
+    corresponding values are passed.  We also assume |m|≤ℓ and |m'|≤ℓ.  Neither of
+    these are checked.  The wedge array that this function indexes is ordered as
+
+        [
+            (ℓ, mp, m) for ℓ in range(ℓₘₐₓ+1)
+            for mp in range(-mpₘₐₓ, mpₘₐₓ+1)
+            for m in range(abs(mp), n+1)
+        ]
+
+    See the docstring of `wedge_index` for
+
+    """
+    if ℓₘₐₓ < 0:
+        return 0
+    if mpₘₐₓ is None or mpₘₐₓ > ℓₘₐₓ:
+        return (ℓₘₐₓ+1) * (ℓₘₐₓ+2) * (2*ℓₘₐₓ+3) // 6
+    else:
+        return ((ℓₘₐₓ+1)*(ℓₘₐₓ+2)*(2*ℓₘₐₓ+3) - 2*(ℓₘₐₓ-mpₘₐₓ)*(ℓₘₐₓ-mpₘₐₓ+1)*(ℓₘₐₓ-mpₘₐₓ+2)) // 6
+
+
+@jit
+def _wedge_index(ℓ, mp, m, mpₘₐₓ):
+    """Helper function for `wedge_index`"""
+    i = wedge_size(ℓ-1, mpₘₐₓ)  # total size of everything with smaller ℓ
+    if mp<1:
+        i += (mpₘₐₓ + mp) * (2*ℓ - mpₘₐₓ + mp + 1) // 2  # size of wedge to the left of m'
+    else:
+        i += (mpₘₐₓ + 1) * (2*ℓ - mpₘₐₓ + 2) // 2  # size of entire left half of wedge
+        i += (mp - 1) * (2*ℓ - mp + 2) // 2  # size of right half of wedge to the left of m'
+    i += m - abs(mp)  # size of column in wedge between m and |m'|
+    return i
+
+
+@jit
+def wedge_index(ℓ, mp, m, mpₘₐₓ=None):
     """Index to "wedge" arrays
 
-    Here, it is assumed that only data with m≥|m'| are stored, and only corresponding values are
-    passed.  We also assume |m|≤ℓ and |m'|≤ℓ.  Neither of these are checked.  The wedge array that
-    this function indexes is ordered as
+    Parameters
+    ----------
+    ℓ : int
+    mp : int
+    m : int
+    mpₘₐₓ : int, optional
+        If None, it is assumed to be at least ℓ
 
-        [(n, mp, m) for n in range(n_max+1) for mp in range(-n, n+1) for m in range(abs(mp), n+1)]
+    See Also
+    --------
+    wedge_size : Total size of wedge array
 
-    """
-    # if m < -mp:
-    #     if m < mp:
-    #         return wedge_index(ℓ, -mp, -m)
-    #     else:
-    #         return wedge_index(ℓ, -m, -mp)
-    # else:
-    #     if m < mp:
-    #         return wedge_index(ℓ, m, mp)
-    if mp>=0:
-        return (2*(ℓ)*(ℓ+1)*(ℓ+2) + 3*mp*(2*ℓ-mp+1) + 6*m) // 6
-    else:
-        return (2*(ℓ)*(ℓ+1)*(ℓ+2) + 3*mp*(2*ℓ+mp+3) + 6*m) // 6
+    Notes
+    -----
+    Here, it is assumed that only data with m≥|m'| are stored, and only
+    corresponding values are passed.  We also assume |m|≤ℓ and |m'|≤ℓ.  Neither of
+    these are checked.  The wedge array that this function indexes is ordered as
 
-
-@jit
-def wedgeify_index(ℓ, mp, m):
-    """Convert general (ℓ, m′, m) index into index appropriate for H wedge
-
-    The wedge is stored as
-
-        [(n, mp, m) for n in range(n_max+1) for mp in range(-n, n+1) for m in range(abs(mp), n+1)]
-
-    This function returns an (ℓ, m′, m) taking advantage of the H symmetries to be the corresponding
-    index lying inside the wedge.
-
-    Obtaining values of the H wedge from other indices:
-    ================================
-    m\m'  -3  -2  -1   0   1   2   3
-     3     v   v   v   v   v   v   v
-     2    nt   v   v   v   v   v   t
-     1    nt  nt   v   v   v   t   t
-     0    nt  nt  nt   v   t   t   t
-    -1    nt  nt   n   n   t   t   t
-    -2    nt   n   n   n   n   t   t
-    -3     n   n   n   n   n   n   t
-    ================================
-    v = value
-    t = transpose, (m',m) -> (m,m')
-    n = negative, (m',m) -> (-m',-m)
-    nt = negative transpose, (m',m) -> (-m,-m')
-
-    Note that the boundaries between the lower wedges are overlap.  For example, the common border
-    of the `n` and `nt` wedges could use either negative or negative transpose and obtain the same
-    element of the main wedge.
+        [
+            (ℓ, mp, m) for ℓ in range(ℓₘₐₓ+1)
+            for mp in range(-mpₘₐₓ, mpₘₐₓ+1)
+            for m in range(abs(mp), n+1)
+        ]
 
     """
+    mp_max = ℓ
+    if mpₘₐₓ is not None:
+        mp_max = min(mpₘₐₓ, mp_max)
     if m < -mp:
         if m < mp:
-            return [ℓ, -mp, -m]
+            return _wedge_index(ℓ, -mp, -m, mp_max)
         else:
-            return [ℓ, -m, -mp]
+            return _wedge_index(ℓ, -m, -mp, mp_max)
     else:
         if m < mp:
-            return [ℓ, m, mp]
+            return _wedge_index(ℓ, m, mp, mp_max)
         else:
-            return [ℓ, mp, m]
+            return _wedge_index(ℓ, mp, m, mp_max)
 
 
 @jit
 def sign(m):
+    """Return sign of input, with sign(0)=1"""
     if m >= 0:
         return 1
     else:
@@ -138,8 +152,14 @@ def sign(m):
 @jit
 def nm_index(n, m):
     """Return flat index into arrray of [n, m] pairs.
-    
-    Assumes array is ordered as [[n, m] for n in range(n_max+1) for m in range(-n, n+1)]
+
+    Assumes array is ordered as
+
+        [
+            [n, m]
+            for n in range(n_max+1)
+            for m in range(-n, n+1)
+        ]
 
     """
     return m + n * (n + 1)
@@ -148,8 +168,14 @@ def nm_index(n, m):
 @jit
 def nabsm_index(n, absm):
     """Return flat index into arrray of [n, abs(m)] pairs
-    
-    Assumes array is ordered as [[n, m] for n in range(n_max+1) for m in range(n+1)]
+
+    Assumes array is ordered as
+
+        [
+            [n, m]
+            for n in range(n_max+1)
+            for m in range(n+1)
+        ]
 
     """
     return absm + (n * (n + 1)) // 2
@@ -158,10 +184,15 @@ def nabsm_index(n, absm):
 @jit
 def nmpm_index(n, mp, m):
     """Return flat index into arrray of [n, mp, m]
-    
+
     Assumes array is ordered as
 
-        [[n, mp, m] for n in range(n_max+1) for mp in range(-n, n+1) for m in range(-n, n+1)]
+        [
+            [n, mp, m]
+            for n in range(n_max+1)
+            for mp in range(-n, n+1)
+            for m in range(-n, n+1)
+        ]
 
     """
     return (((4 * n + 6) * n + 6 * mp + 5) * n + 3 * (m + mp)) // 3
@@ -174,11 +205,11 @@ def _step_1(Hwedge):
 
 
 @jit
-def _step_2(g, h, n_max, Hwedge, Hextra, Hv, cosβ, sinβ):
+def _step_2(g, h, n_max, mp_max, Hwedge, Hextra, Hv, cosβ, sinβ):
     """Compute values H^{0,m}_{n}(β)for m=0,...,n and H^{0,m}_{n+1}(β) for m=0,...,n+1 using Eq. (32):
 
-        H^{0,m}_{n}(β) = (-1)^m \sqrt{(n-|m|)! / (n+|m|)!} P^{|m|}_{n}(cos β)
-                       = (-1)^m (sin β)^m \hat{P}^{|m|}_{n}(cos β) / \sqrt{k (2n+1)}
+        H^{0,m}_{n}(β) = (-1)^m √((n-|m|)! / (n+|m|)!) P^{|m|}_{n}(cos β)
+                       = (-1)^m (sin β)^m P̂^{|m|}_{n}(cos β) / √(k (2n+1))
 
     This function computes the associated Legendre functions directly by recursion
     as explained by Holmes and Featherstone (2002), doi:10.1007/s00190-002-0216-2.
@@ -193,19 +224,19 @@ def _step_2(g, h, n_max, Hwedge, Hextra, Hv, cosβ, sinβ):
     """
     prefactor = np.empty_like(sinβ)
     # n = 1
-    n0n_index = wedge_index(1, 0, 1)
+    n0n_index = wedge_index(1, 0, 1, mp_max)
     nn_index = nm_index(1, 1)
     Hwedge[n0n_index, :] = sqrt3  # Un-normalized
-    Hwedge[n0n_index-1, :] = (g[nn_index-1] * cosβ) / sqrt2  # Normalized
+    Hwedge[n0n_index-1, :] = (g[nn_index-1] * cosβ) * inverse_sqrt2  # Normalized
     # n = 2, ..., n_max+1
     for n in range(2, n_max+2):
         if n <= n_max:
-            n0n_index = wedge_index(n, 0, n)
+            n0n_index = wedge_index(n, 0, n, mp_max)
             H = Hwedge
         else:
             n0n_index = n
             H = Hextra
-        nm10nm1_index = wedge_index(n-1, 0, n-1)
+        nm10nm1_index = wedge_index(n-1, 0, n-1, mp_max)
         nn_index = nm_index(n, n)
         const = np.sqrt(1.0 + 0.5/n)
         gi = g[nn_index-1]
@@ -221,61 +252,63 @@ def _step_2(g, h, n_max, Hwedge, Hextra, Hv, cosβ, sinβ):
             for j in range(H.shape[1]):
                 H[n0n_index-i, j] = gi * cosβ[j] * H[n0n_index-i+1, j] - hi * sinβ[j]**2 * H[n0n_index-i+2, j]
         # if n==2:
-        #     print("args1 =", (g, h, n_max, Hwedge, Hextra, Hv, cosβ, sinβ))
+        #     print("args2 =", (g, h, n_max, Hwedge, Hextra, Hv, cosβ, sinβ))
         # m = 0, with normalization
-        const = np.sqrt(4*n+2)
+        const = 1.0 / np.sqrt(4*n+2)
         gi = g[nn_index-n]
         hi = h[nn_index-n]
         for j in range(H.shape[1]):
-            H[n0n_index-n, j] = (gi * cosβ[j] * H[n0n_index-n+1, j] - hi * sinβ[j]**2 * H[n0n_index-n+2, j]) / const
+            H[n0n_index-n, j] = (gi * cosβ[j] * H[n0n_index-n+1, j] - hi * sinβ[j]**2 * H[n0n_index-n+2, j]) * const
         # Now, loop back through, correcting the normalization for this row, except for n=n element
-        prefactor[:] = 1/const
+        prefactor[:] = const
         for i in range(1, n):
             prefactor *= sinβ
             H[n0n_index-n+i, :] *= prefactor
         # Supply extra edge cases as noted in docstring
         if n <= n_max:
-            Hv[nm_index(n, 1), :] = Hwedge[wedge_index(n, 0, 1)]
-            Hv[nm_index(n, 0), :] = Hwedge[wedge_index(n, 0, 1)]
+            Hv[nm_index(n, 1), :] = Hwedge[wedge_index(n, 0, 1, mp_max)]
+            Hv[nm_index(n, 0), :] = Hwedge[wedge_index(n, 0, 1, mp_max)]
     # Correct normalization of m=n elements
     prefactor[:] = 1.0
     for n in range(1, n_max+1):
         prefactor *= sinβ
-        Hwedge[wedge_index(n, 0, n), :] *= prefactor / np.sqrt(4*n+2)
+        Hwedge[wedge_index(n, 0, n, mp_max), :] *= prefactor / np.sqrt(4*n+2)
     for n in [n_max+1]:
         prefactor *= sinβ
         Hextra[n, :] *= prefactor / np.sqrt(4*n+2)
     # Supply extra edge cases as noted in docstring
-    Hv[nm_index(1, 1), :] = Hwedge[wedge_index(1, 0, 1)]
-    Hv[nm_index(1, 0), :] = Hwedge[wedge_index(1, 0, 1)]
+    Hv[nm_index(1, 1), :] = Hwedge[wedge_index(1, 0, 1, mp_max)]
+    Hv[nm_index(1, 0), :] = Hwedge[wedge_index(1, 0, 1, mp_max)]
 
 
 @jit
-def _step_3(a, b, n_max, Hwedge, Hextra, cosβ, sinβ):
-    """Use relation (41) to compute H^{1,m}_{n}(β) for m=1,...,n.  Using symmetry and shift of the
-    indices this relation can be written as
-        b^{0}_{n+1} H^{1, m}_{n} =   \frac{b^{−m−1}_{n+1} (1−cosβ)}{2} H^{0, m+1}_{n+1}
-                                   − \frac{b^{ m−1}_{n+1} (1+cosβ)}{2} H^{0, m−1}_{n+1}
+def _step_3(a, b, n_max, mp_max, Hwedge, Hextra, cosβ, sinβ):
+    """Use relation (41) to compute H^{1,m}_{n}(β) for m=1,...,n.  Using symmetry and shift
+    of the indices this relation can be written as
+
+        b^{0}_{n+1} H^{1, m}_{n} =   (b^{−m−1}_{n+1} (1−cosβ))/2 H^{0, m+1}_{n+1}
+                                   − (b^{ m−1}_{n+1} (1+cosβ))/2 H^{0, m−1}_{n+1}
                                    − a^{m}_{n} sinβ H^{0, m}_{n+1}
+
     """
     for n in range(1, n_max+1):
         # m = 1, ..., n
-        i1 = wedge_index(n, 1, 1)
+        i1 = wedge_index(n, 1, 1, mp_max)
         if n+1 <= n_max:
-            i2 = wedge_index(n+1, 0, 0)
+            i2 = wedge_index(n+1, 0, 0, mp_max)
             H2 = Hwedge
         else:
             i2 = 0
             H2 = Hextra
         i3 = nm_index(n+1, 0)
         i4 = nabsm_index(n, 1)
-        b5 = b[i3]
+        inverse_b5 = 1.0 / b[i3]
         for i in range(n):
             b6 = b[-i+i3-2]
             b7 = b[i+i3]
             a8 = a[i+i4]
             for j in range(Hwedge.shape[1]):
-                Hwedge[i+i1, j] = (1 / b5) * (
+                Hwedge[i+i1, j] = inverse_b5 * (
                     0.5 * (
                           b6 * (1-cosβ[j]) * H2[i+i2+2, j]
                         - b7 * (1+cosβ[j]) * H2[i+i2, j]
@@ -285,7 +318,7 @@ def _step_3(a, b, n_max, Hwedge, Hextra, cosβ, sinβ):
 
 
 #@jit
-def _step_4(d, n_max, Hwedge, Hv):
+def _step_4(d, n_max, mp_max, Hwedge, Hv):
     """Recursively compute H^{m'+1, m}_{n}(β) for m'=1,...,n−1, m=m',...,n using relation (50) resolved
     with respect to H^{m'+1, m}_{n}:
 
@@ -299,21 +332,21 @@ def _step_4(d, n_max, Hwedge, Hv):
     for n in range(2, n_max+1):
         for mp in range(1, n):
             # m = m', ..., n-1
-            # i1 = wedge_index(n, mp+1, mp)
-            i1 = wedge_index(n, mp+1, mp+1) - 1
-            i2 = wedge_index(n, mp-1, mp)
-            # i3 = wedge_index(n, mp, mp-1)
-            i3 = wedge_index(n, mp, mp) - 1
-            i4 = wedge_index(n, mp, mp+1)
+            # i1 = wedge_index(n, mp+1, mp, mp_max)
+            i1 = wedge_index(n, mp+1, mp+1, mp_max) - 1
+            i2 = wedge_index(n, mp-1, mp, mp_max)
+            # i3 = wedge_index(n, mp, mp-1, mp_max)
+            i3 = wedge_index(n, mp, mp, mp_max) - 1
+            i4 = wedge_index(n, mp, mp+1, mp_max)
             i5 = nm_index(n, mp)
             i6 = nm_index(n, mp-1)
-            d5 = d[i5]
+            inverse_d5 = 1.0 / d[i5]
             d6 = d[i6]
             for i in [0]:
                 d7 = d[i+i6]
                 d8 = d[i+i5]
                 for j in range(Hwedge.shape[1]):
-                    Hv[i+nm_index(n, mp+1), j] = (1 / d5) * (
+                    Hv[i+nm_index(n, mp+1), j] = inverse_d5 * (
                           d6 * Hwedge[i+i2, j]
                         - d7 * Hv[i+nm_index(n, mp), j]
                         + d8 * Hwedge[i+i4, j]
@@ -322,65 +355,67 @@ def _step_4(d, n_max, Hwedge, Hv):
                 d7 = d[i+i6]
                 d8 = d[i+i5]
                 for j in range(Hwedge.shape[1]):
-                    Hwedge[i+i1, j] = (1 / d5) * (
+                    Hwedge[i+i1, j] = inverse_d5 * (
                           d6 * Hwedge[i+i2, j]
                         - d7 * Hwedge[i+i3, j]
                         + d8 * Hwedge[i+i4, j]
                     )
             # if n == 2 and mp == 1:
-            #     print("args1 =", (d, n_max, Hwedge, Hv, i1, i2, i3, i4))
+            #     print("args2 =", (d, n_max, Hwedge, Hv, i1, i2, i3, i4))
             # m = n
             for i in [n-mp]:
                 for j in range(Hwedge.shape[1]):
-                    Hwedge[i+i1, j] = (1 / d5) * (
+                    Hwedge[i+i1, j] = inverse_d5 * (
                           d6 * Hwedge[i+i2, j]
                         - d[i+i6] * Hwedge[i+i3, j]
                     )
 
 
 @jit
-def _step_5(d, n_max, Hwedge, Hv):
+def _step_5(d, n_max, mp_max, Hwedge, Hv):
     """Recursively compute H^{m'−1, m}_{n}(β) for m'=−1,...,−n+1, m=−m',...,n using relation (50)
     resolved with respect to H^{m'−1, m}_{n}:
+
       d^{m'−1}_{n} H^{m'−1, m}_{n} = d^{m'}_{n} H^{m'+1, m}_{n}
                                      + d^{m−1}_{n} H^{m', m−1}_{n}
                                      − d^{m}_{n} H^{m', m+1}_{n}
+
     (where the last term drops out for m=n).
 
-    NOTE: Although arxiv:1403.7698 specifies the loop over mp to start at -1, I find it necessary to
-    start at 0, or there will be missing information.  This also requires setting the (m',m)=(0,-1)
-    components before beginning this loop.
+    NOTE: Although arxiv:1403.7698 specifies the loop over mp to start at -1, I
+    find it necessary to start at 0, or there will be missing information.  This
+    also requires setting the (m',m)=(0,-1) components before beginning this loop.
 
     """
     for n in range(0, n_max+1):
         for mp in range(0, -n, -1):
             # m = -m', ..., n-1
-            # i1 = wedge_index(n, mp-1, -mp)
-            i1 = wedge_index(n, mp-1, -mp+1) - 1
-            # i2 = wedge_index(n, mp+1, -mp)
-            i2 = wedge_index(n, mp+1, -mp+1) - 1
-            # i3 = wedge_index(n, mp, -mp-1)
-            i3 = wedge_index(n, mp, -mp) - 1
-            i4 = wedge_index(n, mp, -mp+1)
+            # i1 = wedge_index(n, mp-1, -mp, mp_max)
+            i1 = wedge_index(n, mp-1, -mp+1, mp_max) - 1
+            # i2 = wedge_index(n, mp+1, -mp, mp_max)
+            i2 = wedge_index(n, mp+1, -mp+1, mp_max) - 1
+            # i3 = wedge_index(n, mp, -mp-1, mp_max)
+            i3 = wedge_index(n, mp, -mp, mp_max) - 1
+            i4 = wedge_index(n, mp, -mp+1, mp_max)
             i5 = nm_index(n, mp-1)
             i6 = nm_index(n, mp)
             i7 = nm_index(n, -mp-1)
             i8 = nm_index(n, -mp)
-            d5 = d[i5]
+            inverse_d5 = 1.0 / d[i5]
             d6 = d[i6]
             for i in [0]:
                 d7 = d[i+i7]
                 d8 = d[i+i8]
                 if mp == 0:
                     for j in range(Hwedge.shape[1]):
-                        Hv[i+nm_index(n, mp-1), j] = (1 / d5) * (
+                        Hv[i+nm_index(n, mp-1), j] = inverse_d5 * (
                               d6 * Hv[i+nm_index(n, mp+1), j]
                             + d7 * Hv[i+nm_index(n, mp), j]
                             - d8 * Hwedge[i+i4, j]
                         )
                 else:
                     for j in range(Hwedge.shape[1]):
-                        Hv[i+nm_index(n, mp-1), j] = (1 / d5) * (
+                        Hv[i+nm_index(n, mp-1), j] = inverse_d5 * (
                               d6 * Hwedge[i+i2, j]
                             + d7 * Hv[i+nm_index(n, mp), j]
                             - d8 * Hwedge[i+i4, j]
@@ -389,7 +424,7 @@ def _step_5(d, n_max, Hwedge, Hv):
                 d7 = d[i+i7]
                 d8 = d[i+i8]
                 for j in range(Hwedge.shape[1]):
-                    Hwedge[i+i1, j] = (1 / d5) * (
+                    Hwedge[i+i1, j] = inverse_d5 * (
                           d6 * Hwedge[i+i2, j]
                         + d7 * Hwedge[i+i3, j]
                         - d8 * Hwedge[i+i4, j]
@@ -397,52 +432,37 @@ def _step_5(d, n_max, Hwedge, Hv):
             # m = n
             i = n+mp
             for j in range(Hwedge.shape[1]):
-                Hwedge[i+i1, j] = (1 / d5) * (
+                Hwedge[i+i1, j] = inverse_d5 * (
                       d6 * Hwedge[i+i2, j]
                     + d[i+i7] * Hwedge[i+i3, j]
                 )
 
 
 class HCalculator(object):
-    def __init__(self, n_max):
+    def __init__(self, n_max, mp_max=None):
         """Object to repeatedly calculate Wigner H values
 
-        The H matrix is related to Wigner's d matrix as given by arxiv:1403.7698 according to
+        In particular the simplicity of the first two of these relations implies that
+        we only need to compute one fourth of the total number of elements.  There is
+        also a very accurate and efficient recursion method to compute these values.
 
-        dₗⁿᵐ = ϵₙ ϵ₋ₘ Hₗⁿᵐ, where
+        Create this object using the largest value of `n` (also commonly denoted `j` or
+        `ℓ`) you expect to need, optionally create the `workspace` for a given shape of
+        cosβ using that method, and then call this object for a given value or array of
+        cosβ values.
 
-               ⎧ 1 for k≤0
-          ϵₖ = ⎨
-               ⎩ (-1)ᵏ for k>0
-
-        This object has various advantageous features compared to the d matrix.  It obeys these
-        symmetry relations:
-
-            H^{m', m}_n(\beta) = H^{m, m'}_n(\beta)
-            H^{m', m}_n(\beta) = H^{-m', -m}_n(\beta)
-            H^{m', m}_n(\beta) = (-1)^{n+m+m'} H^{-m', m}_n(\pi - \beta)
-            H^{m', m}_n(\beta) = (-1)^{m+m'} H^{m', m}_n(-\beta)
-
-        In particular the simplicity of the first two of these relations implies that we only need
-        to compute one fourth of the total number of elements.  There is also a very accurate and
-        efficient recursion method to compute these values.
-
-        Create this object using the largest value of `n` (also commonly denoted `j` or `ℓ`) you
-        expect to need, optionally create the `workspace` for a given shape of cosβ using that method,
-        and then call this object for a given value or array of cosβ values.
-
-        The returned object is a series of "wedges" of the matrix, for the various values of `n`,
-        comprising just a quarter of the elements of the full matrix; all remaining elements are
-        determined by the first two symmetries above.  This wedge has an initial dimension
-        representing a multi-index for (ℓ, mp, m) values, while following dimensions are just the
-        same as cosβ.  Any required value may be obtained with `wedge_index(*wedgeify_index(ℓ, mp,
-        m))`.  The inner call translates a general index into the equivalent index lying inside the
-        wedge, while the outer call translates that corrected (ℓ, mp, m) tuple into a linear index
-        into the array.
+        The returned object is a series of "wedges" of the matrix, for the various
+        values of `n`, comprising just a quarter of the elements of the full matrix;
+        all remaining elements are determined by the first two symmetries above.  This
+        wedge has an initial dimension representing a multi-index for (ℓ, mp, m)
+        values, while following dimensions are just the same as cosβ.  Any required
+        value may be obtained with `wedge_index(ℓ, mp, m, mp_max)`.  The inner call
+        translates a general index into the equivalent index lying inside the wedge,
+        while the outer call translates that corrected (ℓ, mp, m) tuple into a linear
+        index into the array.
 
         Example Usage
-        =============
-
+        -------------
         hcalc = HCalculator(n_max)
         workspace = hcalc.workspace(cosβ)  # Note that cosβ can be an array of many values
         # Possibly loop over many values of cosβ here
@@ -450,9 +470,13 @@ class HCalculator(object):
 
         """
         self.n_max = int(n_max)
+        if mp_max is not None:
+            self.mp_max = abs(int(mp_max))
+        else:
+            self.mp_max = None
         if self.n_max < 0:
             raise ValueError('Nonsensical value for n_max = {0}'.format(self.n_max))
-        self.wedge_size = wedge_size(self.n_max)
+        self.wedge_size = wedge_size(self.n_max, self.mp_max)
         n = np.array([n for n in range(self.n_max+2) for m in range(-n, n+1)])
         m = np.array([m for n in range(self.n_max+2) for m in range(-n, n+1)])
         absn = np.array([n for n in range(self.n_max+2) for m in range(n+1)])
@@ -472,7 +496,7 @@ class HCalculator(object):
         ):
             raise ValueError("Found a non-finite value inside this object")
 
-    def workspace(self, cosβ=1.0):
+    def workspace(self, cosβ=[1.0]):
         """Return a new workspace sized for cosβ."""
         cosβ = np.asarray(cosβ, dtype=float)
         return np.zeros((self.wedge_size+(self.n_max+1)**2+self.n_max+2,) + cosβ.shape, dtype=float)
@@ -496,24 +520,15 @@ class HCalculator(object):
         Hv = workspace[self.wedge_size:self.wedge_size+(self.n_max+1)**2]
         Hextra = workspace[self.wedge_size+(self.n_max+1)**2:self.wedge_size+(self.n_max+1)**2+self.n_max+2]
         _step_1(Hwedge)
-        _step_2(self.g, self.h, self.n_max, Hwedge, Hextra, Hv, cosβ, sinβ)
-        _step_3(self.a, self.b, self.n_max, Hwedge, Hextra, cosβ, sinβ)
-        _step_4(self.d, self.n_max, Hwedge, Hv)
-        _step_5(self.d, self.n_max, Hwedge, Hv)
+        _step_2(self.g, self.h, self.n_max, self.mp_max, Hwedge, Hextra, Hv, cosβ, sinβ)
+        _step_3(self.a, self.b, self.n_max, self.mp_max, Hwedge, Hextra, cosβ, sinβ)
+        _step_4(self.d, self.n_max, self.mp_max, Hwedge, Hv)
+        _step_5(self.d, self.n_max, self.mp_max, Hwedge, Hv)
         Hwedge.reshape((-1,)+cosβshape)
         return Hwedge
 
     def wigner_d(self, cosβ, sinβ=None, workspace=None):
         """Return Wigner's d matrix"""
-        from .. import LMpM_total_size, LMpM_index
-        from .wignerH2 import wedge_index as wedge_index2
-        def ϵ(m):
-            if m <= 0:
-                return 1
-            elif m%2:
-                return -1
-            else:
-                return 1
         ell_min = 0
         ell_max = self.n_max
         Hwedge = self(cosβ, sinβ, workspace)
@@ -522,6 +537,6 @@ class HCalculator(object):
             for mp in range(-ell, ell+1):
                 for m in range(-ell, ell+1):
                     i_d = LMpM_index(ell, mp, m, ell_min)
-                    i_H = wedge_index2(ell, mp, m)#, self.mp_max)
+                    i_H = wedge_index(ell, mp, m, self.mp_max)
                     d[i_d] = ϵ(mp) * ϵ(-m) * Hwedge[i_H]
         return d
